src/requirements.txt
openai>=1.0.0
pyyaml>=6.0
rank-bm25>=0.2.2
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
transformers>=4.45.0
torch>=2.2.0
sentence-transformers>=2.7.0
datasets>=2.18.0
accelerate>=0.26.0
